Clean up field emitter mess
