--- src/FlatSharp.Compiler/Emitters/IFieldDefinitionEmitter.cs
@@ -1,21 +1,0 @@
-﻿namespace FlatSharp.Compiler
-{
-    internal interface IFieldDefinitionEmitter
-    {
-        bool EmitDefinition(
-            CodeWriter writer,
-            FieldDefinition fieldDefinition,
-            CompileContext context);
-
-        void EmitCloneLine(
-            CodeWriter writer,
-            string variableName,
-            FieldDefinition fieldDefinition,
-            CompileContext compileContext);
-
-        void EmitDefaultInitializationLine(
-            CodeWriter writer,
-            FieldDefinition fieldDefinition,
-            CompileContext compileContext);
-    }
-}
